requirement.txt
torch==2.3.1 
torchvision==0.18.1 
torchaudio==2.3.1
transformers==4.42.3 
peft==0.11.1
torchlibrosa==0.1.0
loguru==0.7.2 
einops==0.8.0 
timm==1.0.7
torch-audiomentations==0.11.1
sentencepiece==0.2.0
protobuf==5.27.2